Hidtmgen conversion + gtests (#4137)
* Initial hidtmgen conversion + gtests

* Initial hidtmgen tests
