+++ isis/src/mro/apps/hidtmgen/hidtmgen.h
@@ -1,0 +1,11 @@
+#ifndef hidtmgen_h // Change this to your app name in all lower case suffixed with _h (e.g. campt_h, cam2map_h etc.)
+#define hidtmgen_h
+
+#include "Cube.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void hidtmgen(UserInterface &ui);
+}
+
+#endif
