+++ isis/tests/data/hidtmgen/color/sequenceNumbers.txt
@@ -1,0 +1,2 @@
+31
+41
