+++ isis/tests/data/hidtmgen/equi/orthoInputList.txt
@@ -1,0 +1,2 @@
+data/hidtmgen/ortho/PSP_002118_1510_1m_o_forPDS_cropped.cub
+data/hidtmgen/ortho/PSP_002118_1510_25cm_o_forPDS_cropped.cub
