+++ isis/tests/data/hidtmgen/equi/params.pvl
@@ -1,0 +1,11 @@
+DATA_SET_ID               = "MRO-M-HIRISE-5-DTM-V1.0"
+DATA_SET_NAME             = "MRO MARS HIGH RESOLUTION IMAGING SCIENCE EXPERIMENT DTM V1.0"
+PRODUCER_INSTITUTION_NAME = "UNIVERSITY OF ARIZONA"
+PRODUCER_ID               = "UA"
+PRODUCER_FULL_NAME        = "ALFRED MCEWEN"
+PRODUCT_VERSION_ID        = "2.0"
+SOFTWARE_NAME             = "Socet_Set 5.4.1"
+PRODUCING_INSTITUTION     = "A"
+RATIONALE_DESC            = "Equi test"
+DTM_SOURCE_PRODUCT_ID     = (PSP_002118_1510, PSP_003608_1510)
+END
