+++ isis/tests/data/hidtmgen/equi/sequenceNumbers.txt
@@ -1,0 +1,2 @@
+01
+02
