+++ isis/tests/data/hidtmgen/error/orthoInputList2Bands.txt
@@ -1,0 +1,1 @@
+data/hidtmgen/ortho/2BandImage.cub
