+++ isis/tests/data/hidtmgen/error/orthoOutputFiles.lis
@@ -1,0 +1,2 @@
+ORTHO1.IMG
+ORTHO2.IMG
