+++ isis/tests/data/hidtmgen/error/orthoProdList3Item.txt
@@ -1,0 +1,3 @@
+Ortho1Product
+Ortho2Product
+Ortho3Product
