+++ isis/tests/data/hidtmgen/error/sequenceNumbers2item.txt
@@ -1,0 +1,2 @@
+31
+41
