+++ isis/tests/data/hidtmgen/nonDefaultNames/orthoOutputProductIds.lis
@@ -1,0 +1,2 @@
+Ortho1Product
+Ortho2Product
