+++ isis/tests/data/hidtmgen/nonDefaultNames/params.pvl
@@ -1,0 +1,11 @@
+DATA_SET_ID               = "DataSetIdFromParamsPvl"
+DATA_SET_NAME             = "Data Set Name From PARAMSPVL"
+PRODUCER_INSTITUTION_NAME = "Producer Institution Name from PARAMSPVL"
+PRODUCER_ID               = "Producer Id from PARAMSPVL"
+PRODUCER_FULL_NAME        = "Producer Full Name from PARAMSPVL"
+PRODUCING_INSTITUTION     = "Z"
+PRODUCT_VERSION_ID        = "3.14"
+RATIONALE_DESC            = "Test DefaultNames=false"
+SOFTWARE_NAME             = "Software Name from PARAMSPVL"
+DTM_SOURCE_PRODUCT_ID     = (PSP_002118_1510, PSP_003608_1510)
+END
