+++ isis/tests/data/hidtmgen/orthoOnly/params.pvl
@@ -1,0 +1,14 @@
+# See hidtmgen documentation for explanation of 
+# DTM_SOURCE_PRODUCT_ID and ORTHO_SOURCE_DTM_ID codes
+DATA_SET_ID               = "MRO-M-HIRISE-5-DTM-V1.0"
+DATA_SET_NAME             = "MRO MARS HIGH RESOLUTION IMAGING SCIENCE EXPERIMENT DTM V1.0"
+PRODUCER_INSTITUTION_NAME = "UNIVERSITY OF ARIZONA"
+PRODUCER_ID               = "UA"
+PRODUCER_FULL_NAME        = "ALFRED MCEWEN"
+PRODUCT_VERSION_ID        = "0.042"
+SOFTWARE_NAME             = "Socet_Set 5.4.1"
+PRODUCING_INSTITUTION     = "A"
+RATIONALE_DESC            = ""
+DTM_SOURCE_PRODUCT_ID    = (pSP_xxxxxx_xxxx, pSP_yyyyyy_yyyy)
+ORTHO_SOURCE_DTM_ID      = DTems_xxxxxx_xxxx_yyyyyy_yyyy_vnn
+END
