+++ isis/tests/data/hidtmgen/orthoOnly/sequenceNumbers.txt
@@ -1,0 +1,2 @@
+31
+41
