+++ isis/tests/data/hidtmgen/outputTypes/params.pvl
@@ -1,0 +1,12 @@
+DATA_SET_ID               = "MRO-M-HIRISE-5-DTM-V1.0"
+DATA_SET_NAME             = "MRO MARS HIGH RESOLUTION IMAGING SCIENCE EXPERIMENT DTM V1.0"
+PRODUCER_INSTITUTION_NAME = "UNIVERSITY OF ARIZONA"
+PRODUCER_ID               = "UA"
+PRODUCER_FULL_NAME        = "ALFRED MCEWEN"
+PRODUCT_VERSION_ID        = "3.14"
+SOFTWARE_NAME             = "Socet_Set 5.4.1"
+PRODUCING_INSTITUTION     = "A"
+RATIONALE_DESC            = "Testing hidtmgen byte order, special pixels, bit type, and version ID"
+DTM_SOURCE_PRODUCT_ID     = (ESP_042252_1930, ESP_042753_1930)
+ORTHO_SOURCE_DTM_ID       = dtmID
+END
