+++ isis/tests/data/hidtmgen/outputTypes/sequenceNumbers.txt
@@ -1,0 +1,2 @@
+31
+41
