+++ isis/tests/data/hidtmgen/polar/orthosequencenumberlist.txt
@@ -1,0 +1,2 @@
+1
+2
