+++ isis/tests/data/hidtmgen/polar/params.pvl
@@ -1,0 +1,11 @@
+DATA_SET_ID               = "DataSetIdFromParamsPvl"
+DATA_SET_NAME             = "Data Set Name From PARAMSPVL"
+PRODUCER_INSTITUTION_NAME = "Producer Institution Name from PARAMSPVL"
+PRODUCER_ID               = "Producer Id from PARAMSPVL"
+PRODUCER_FULL_NAME        = "Producer Full Name from PARAMSPVL"
+PRODUCING_INSTITUTION     = "Z"
+PRODUCT_VERSION_ID        = "12.0"
+RATIONALE_DESC            = "Polar test"
+SOFTWARE_NAME             = "Software Name from PARAMSPVL"
+DTM_SOURCE_PRODUCT_ID     = (PSP_009404_2635, PSP_010221_2635)
+END
